updated github data access
