+++ augur/tasks/github/util/github_data_access.py
@@ -176,112 +176,3 @@ class GithubDataAccess:
         merged_params = {**original_params, **additional_params}
         updated_query = urlencode(merged_params, doseq=True)
         return url_components._replace(query=updated_query).geturl()
-
-# ---- TESTS ----
-
-if __name__ == "__main__":
-    import pytest
-    import httpx
-    from unittest.mock import MagicMock
-
-    class DummyKeyClient:
-        def __init__(self):
-            self.called = {"request": 0, "expire": 0, "invalidate": 0}
-        def request(self):
-            self.called["request"] += 1
-            return "dummy_key"
-        def expire(self, key, expire_time):
-            self.called["expire"] += 1
-            return "dummy_key_expired"
-        def invalidate(self, key):
-            self.called["invalidate"] += 1
-            return "dummy_key_invalid"
-
-    @pytest.fixture
-    def setup_github_da(monkeypatch):
-        logger = logging.getLogger("github_test")
-        logger.setLevel(logging.DEBUG)
-        key_client = DummyKeyClient()
-        return GithubDataAccess(key_client, logger)
-
-    def make_mock_response(
-        status_code,
-        headers=None,
-        json_data=None,
-        text=None,
-        links=None,
-    ):
-        class DummyResponse:
-            def __init__(self):
-                self.status_code = status_code
-                self.headers = headers or {}
-                self._json = json_data
-                self.text = text or ""
-                self.links = links or {}
-            def json(self):
-                if self._json is not None:
-                    return self._json
-                raise ValueError("No json data")
-            def raise_for_status(self):
-                if 400 <= self.status_code < 600:
-                    raise httpx.HTTPStatusError(
-                        f"{self.status_code} Error", request=None, response=None
-                    )
-        return DummyResponse()
-
-    def test_403_with_rate_limit_message(setup_github_da, monkeypatch):
-        gda = setup_github_da
-        monkeypatch.setattr(gda, "key", "abc")
-        resp = make_mock_response(403, json_data={"message": "API rate limit exceeded for 1.2.3.4"})
-        def client_request(*args, **kwargs):
-            return resp
-        with monkeypatch.context() as m:
-            m.setattr(httpx.Client, "request", client_request)
-            with pytest.raises(RatelimitException):
-                gda.make_request("https://api.github.com/test")
-
-    def test_403_with_secondary_rate_limit(setup_github_da, monkeypatch):
-        gda = setup_github_da
-        monkeypatch.setattr(gda, "key", "abc")
-        resp = make_mock_response(403, json_data={"message": "You have exceeded a secondary rate limit. Please wait."})
-        def client_request(*args, **kwargs):
-            return resp
-        with monkeypatch.context() as m:
-            m.setattr(httpx.Client, "request", client_request)
-            with pytest.raises(RatelimitException):
-                gda.make_request("https://api.github.com/test")
-
-    def test_400_abuse_detection(setup_github_da, monkeypatch):
-        gda = setup_github_da
-        monkeypatch.setattr(gda, "key", "abc")
-        resp = make_mock_response(400, json_data={"message": "You have triggered an abuse detection mechanism."})
-        def client_request(*args, **kwargs):
-            return resp
-        with monkeypatch.context() as m:
-            m.setattr(httpx.Client, "request", client_request)
-            with pytest.raises(RatelimitException):
-                gda.make_request("https://api.github.com/test")
-
-    def test_429_rate_limit(setup_github_da, monkeypatch):
-        gda = setup_github_da
-        monkeypatch.setattr(gda, "key", "abc")
-        resp = make_mock_response(429, headers={"Retry-After": "60"})
-        def client_request(*args, **kwargs):
-            return resp
-        with monkeypatch.context() as m:
-            m.setattr(httpx.Client, "request", client_request)
-            with pytest.raises(RatelimitException):
-                gda.make_request("https://api.github.com/test")
-
-    def test_xratelimit_remaining_low(setup_github_da, monkeypatch):
-        gda = setup_github_da
-        monkeypatch.setattr(gda, "key", "abc")
-        resp = make_mock_response(200, headers={"X-RateLimit-Remaining": "3"})
-        def client_request(*args, **kwargs):
-            return resp
-        with monkeypatch.context() as m:
-            m.setattr(httpx.Client, "request", client_request)
-            with pytest.raises(RatelimitException):
-                gda.make_request("https://api.github.com/test")
-
-    print("All tests passed! (run with pytest for more detail)")
